requirements.txt
tornado
bcrypt
mysql-connector-python